requirements.txt
cassandra-driver>=3.25.0
python-dotenv
pytz